server/requirements.txt
aiohttp==3.9.1
aioice==0.9.0
aiortc==1.6.0
aiosignal==1.3.1
alembic==1.13.1
amqp==5.2.0
aniso8601==9.0.1
apispec==6.3.1
asgiref==3.7.2
attrs==23.2.0
av==11.0.0
bcrypt==4.1.2
bidict==0.22.1
billiard==4.2.0
blinker==1.7.0
celery==5.3.6
cffi==1.16.0
click==8.1.7
click-didyoumean==0.3.0
click-plugins==1.1.1
click-repl==0.3.0
colorama==0.4.6
cryptography==41.0.7
dnspython==2.4.2
Flask==3.0.0
flask-apispec==0.11.4
Flask-Bcrypt==1.0.1
Flask-Celery==2.4.3
Flask-Cors==4.0.0
Flask-JWT-Extended==4.6.0
Flask-Mail==0.9.1
flask-marshmallow==0.15.0
Flask-Migrate==4.0.5
flask-restx==1.3.0
Flask-Script==2.0.6
Flask-SocketIO==5.3.6
Flask-SQLAlchemy==3.1.1
frozenlist==1.4.1
google-crc32c==1.5.0
greenlet==3.0.3
h11==0.14.0
h2==4.1.0
hpack==4.0.0
Hypercorn==0.16.0
hyperframe==6.0.1
idna==3.6
ifaddr==0.2.0
importlib-resources==6.1.1
iniconfig==2.0.0
itsdangerous==2.1.2
Jinja2==3.1.2
jsonschema==4.20.0
jsonschema-specifications==2023.12.1
kombu==5.3.4
Mako==1.3.0
MarkupSafe==2.1.3
marshmallow==3.20.1
multidict==6.0.4
packaging==23.2
pluggy==1.3.0
priority==2.0.0
prompt-toolkit==3.0.43
pycparser==2.21
pyee==11.1.0
PyJWT==2.8.0
pylibsrtp==0.9.1
pyOpenSSL==23.3.0
pytest==7.4.4
python-dateutil==2.8.2
python-engineio==4.8.1
python-socketio==5.10.0
pytz==2023.3.post1
referencing==0.32.0
rpds-py==0.16.2
simple-websocket==1.0.0
six==1.16.0
SQLAlchemy==2.0.25
typing_extensions==4.9.0
tzdata==2023.4
vine==5.1.0
wcwidth==0.2.12
webargs==8.3.0
Werkzeug==3.0.1
wsproto==1.2.0
yarl==1.9.4
